various improvements
